Add bootstrap script
The bootstrap script is all that's needed to get these dotfiles
linked and ready to go on a fresh system. It requires git and rcup
to already be installed to work, which may change in the future,
but this takes care of cloning the repository and performing the
initial link.